Moves models to app dir
