+++ lib/bot.js
@@ -3,8 +3,8 @@
 const fs = require('fs');
 const RiveScript = require('rivescript');
 const mongoose = require('../config/mongoose');
-const Users = require('../models/User');
-const Campaigns = require('../models/Campaign');
+const Users = require('../app/models/User');
+const Campaigns = require('../app/models/Campaign');
 const helpers = require('./helpers');
 const config = require('../config/lib/bot');
